+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.6.1 (2026-08-24)
+
+### Fix
+
+- reject empty arrays in polynomial creation (#144)
+
+## 0.6.0 (2026-08-24)
+
+### Feat
+
+- remove extra variables with squeeze method (#143)
+
 ## 0.5.0 (2026-08-06)
 
 ### Feat